[back] Refactor handlePost into filepond.ts
